ci: also launch tests on i386
Fixes #6
Fixes #46

The tests in conversion_test.go are the ones that are not architecture dependent
The tests in conversion_64bit_test.go complete them for 64-bit systems

The tests in examples_test.go are the ones that are not architecture dependent
The tests in examples_32bit_test.go are for 32-bit systems
The tests in examples_64bit_test.go are for 64-bit systems

The architecture dependent file covers the fact, you can reach a higher value with int and uint
on 64-bit systems, but you will get a compile error on 32-bit.

The error message is different on 32-bit and 64-bit systems
The max is 9223372036854775807 on 64-bit and 2147483647 on 32-bit
