+++ .github/workflows/golang.yml
@@ -33,3 +33,6 @@ jobs:
 
       - name: Launch golangci-lint
         uses: golangci/golangci-lint-action@v6.1.1
+
+      - name: Run tests on i386
+        run: GOARCH=386 go test
